requirements.txt
websocket-client
colorama